Remove duplicate code
Remove duplicate capture method.
