+++ lib/pbench/test/conftest.py
@@ -1,0 +1,9 @@
+import subprocess
+
+import pytest
+
+@pytest.helpers.register
+def capture(command):
+    proc = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+    out, err = proc.communicate()
+    return out, err, proc.returncode
